requirements.txt
asgiref==3.4.1
Django==3.2.13
django-ckeditor==6.3.2
django-cleanup==6.0.0
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==8.4.0
psycopg2-binary==2.9.3
pytz==2022.1
sorl-thumbnail==12.8.0
sqlparse==0.4.2
typing_extensions==4.1.1
